MIN-103 팀 평점 타입 변경

+++ api/src/main/java/com/minecraft/job/api/controller/ReviewApi.java
@@ -21,7 +21,7 @@ public class ReviewApi {
 
     @PostMapping
     public ReviewCreateResponse create(@RequestBody ReviewCreateRequest req) {
-        Pair<Review, Long> pair = reviewAppService.create(req.toDto());
+        Pair<Review, Double> pair = reviewAppService.create(req.toDto());
 
         return ReviewCreateResponse.create(pair);
     }
