+++ common/src/main/java/com/minecraft/job/common/team/service/TeamService.java
@@ -6,7 +6,7 @@ public interface TeamService {
 
     Team create(Long userId, String name, String description, String logo, Long memberNum);
 
-    void applyAveragePoint(Long teamId, Long averagePoint);
+    void applyAveragePoint(Long teamId, Double averagePoint);
 
     void update(Long teamId, Long userId, String name, String description, String logo, Long memberNum);
